+++ src/diepvries/deserializers/snowflake_deserializer.py
@@ -5,9 +5,10 @@ import logging
 from collections import defaultdict
 from dataclasses import asdict, dataclass
 from functools import cached_property
-from typing import Dict, List, Literal, Optional, Type
+from typing import Dict, List, Optional, Type
 
 from snowflake.connector import DictCursor, connect
+from snowflake.connector.network import DEFAULT_AUTHENTICATOR
 
 from .. import TABLE_PREFIXES, FieldDataType, FixedPrefixLoggerAdapter, TableType
 from ..driving_key_field import DrivingKeyField
@@ -32,15 +33,15 @@ class DatabaseConfiguration:
     warehouse: str
     account: str
     password: Optional[str] = None
-    authenticator: Optional[Literal["password", "externalbrowser"]] = "password"
+    authenticator: Optional[str] = DEFAULT_AUTHENTICATOR
 
     def __post_init__(self):
         """Validate input for optional attributes."""
-        if self.authenticator == "password" and not self.password:
+        if self.authenticator == DEFAULT_AUTHENTICATOR and not self.password:
             raise ValueError(
-                "Password was not provided. It is a mandatory attribute when the "
-                "authenticator is `password`. Empty passwords are only allowed "
-                "when `authenticator='externalbrowser'`."
+                f"Password was not provided. It is a mandatory attribute when the "
+                f"authenticator is not `{DEFAULT_AUTHENTICATOR}`. Empty passwords are "
+                f"only allowed when `authenticator='externalbrowser'`."
             )